Locali_API/requirements.txt
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-mysql==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-extensions==0.6.0
drf-flex-fields==0.8.5
drf-yasg==1.17.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
ordereddict==1.1
packaging==20.4
Pillow==7.1.2
protobuf==3.12.1
psycopg2-binary==2.8.5
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.25.8
zope.interface==5.1.0
